django/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment variables
python-decouple>=3.8,<4.0

# Development tools
django-debug-toolbar>=4.0,<5.0

# API documentation
drf-spectacular>=0.26,<1.0

# Production
gunicorn>=21.0,<22.0
whitenoise>=6.5,<7.0

# Testing
pytest>=7.0,<8.0
pytest-django>=4.5,<5.0
factory-boy>=3.2,<4.0

# Code quality
black>=23.0,<24.0
flake8>=6.0,<7.0
isort>=5.12,<6.0 